feat: 优化逻辑

+++ src/utils/myApp.ts
@@ -1,3 +1,4 @@
+import { checkHasInstall, launchApp } from "./adb";
 import { execCmd } from "./index";
 
 class MyAppBridge {
@@ -34,20 +35,36 @@ class MyAppBridge {
   public installApp() {
     return new Promise(async (resolve, reject) => {
       let successList = [];
+      let timeoutFlag = true;
+
+      let timer = setTimeout(() => {
+        if (timeoutFlag) {
+          console.log('App install timeout')
+          reject('App install timeout')
+        }
+      }, 60000)
+
       const installCallback = (index) => {
         successList.push(index);
         if (index.length >= this.devices.length) {
           // 全部安装成功
           console.log('all installed')
+          timeoutFlag = false;
+          clearTimeout(timer);
           resolve(true);
         }
       }
 
       for( let i in this.devices) {
         const item = this.devices[i];
-        const excRsp = await execCmd(`adb -s ${item.name} install ${this.apk}`);
+        const existRsp = await checkHasInstall(this.apkName, item.name);
+        if (existRsp) {
+          console.log(`${this.apkName} has Installed`)
+          return resolve(true);
+        }
+        await execCmd(`adb -s ${item.name} install ${this.apk}`);
         let timer = setInterval(async () => {
-          const excRsp = await execCmd(`adb -s ${item.name} shell pm -l | grep ${this.apkName}`);
+          const excRsp = await checkHasInstall(this.apkName, item.name);
           if (excRsp) {
             // 安装成功
             console.log('install success')
@@ -60,6 +77,12 @@ class MyAppBridge {
     })
   };
 
+  public async startApp() {
+    for (let i in this.devices) {
+      const item = this.devices[i];
+      await launchApp(this.apkName, item.name)
+    }
+  }
 }
 
 export default MyAppBridge;
